misc/HtmlTestSuite/results/test-0496.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><meta charset="utf-8"/>
<title>CSS Grid Layout Test: Mininum height of grid items orthogonal</title>
<link href="mailto:obrufau@igalia.com" rel="author" title="Oriol Brufau"/>
<link href="https://drafts.csswg.org/css-grid-1/#grid-items" rel="help"/>
<meta content="Checks that orthogonal grid items minimum height take into account borders, padding and margins for grid containers with definite height." name="assert"/>
<link href="/css/support/grid.css" rel="stylesheet"/>
<style>
.grid {
  display: inline-grid;
  height: 100px;
  border: solid thick;
  grid: minmax(auto, 0px) / 10px 10px;
}

.grid &gt; div:nth-child(1) { background: cyan; writing-mode: vertical-lr; }
.grid &gt; div:nth-child(2) { background: magenta; }

.height60 { height: 60px; }
.minHeight60 { min-height: 60px; }

.marginTop5 { margin-top: 5px; }
.marginBottom10 { margin-bottom: 10px; }
.marginLeftAuto { margin-left: auto }

.paddingTop6 { padding-top: 6px; }
.paddingBottom3 { padding-bottom: 3px; }

.borderTop2, .borderBottom4 { border: solid yellow 0px; }
.borderTop2 { border-top-width: 2px; }
.borderBottom4 { border-bottom-width: 4px; }
</style>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/resources/check-layout-th.js"/>

</head><body onload="checkLayout('.grid')">

<div id="log"/>

<h3>Direction LTR</h3>

<pre>Item height: 60px;</pre>

<div class="grid">
  <div class="height60" data-expected-height="60"/>
  <div data-expected-height="60"/>
</div>

<pre>Item min-height: 60px;</pre>

<div class="grid">
  <div class="minHeight60" data-expected-height="60"/>
  <div data-expected-height="60"/>
</div>

<pre>Item height: 60px; &amp; margin-left: 5px;</pre>

<div class="grid">
  <div class="height60 marginTop5" data-expected-height="60"/>
  <div data-expected-height="65"/>
</div>

<pre>Item min-height: 60px; &amp; margin-left: 5px;</pre>

<div class="grid">
  <div class="minHeight60 marginTop5" data-expected-height="60"/>
  <div data-expected-height="65"/>
</div>

<pre>Item min-height: 60px; &amp; margin-top: 5px; &amp; margin-left: auto</pre>

<div class="grid">
  <div class="minHeight60 marginTop5 marginLeftAuto" data-expected-height="60"/>
  <div data-expected-height="65"/>
</div>

<pre>Item height: 60px; &amp; margin-right: 10px;</pre>

<div class="grid">
  <div class="height60 marginBottom10" data-expected-height="60"/>
  <div data-expected-height="70"/>
</div>

<pre>Item min-height: 60px; &amp; margin-right: 10px;</pre>

<div class="grid">
  <div class="minHeight60 marginBottom10" data-expected-height="60"/>
  <div data-expected-height="70"/>
</div>

<pre>Item height: 60px; &amp; margin-left: 5px; &amp; margin-right: 10px;</pre>

<div class="grid">
  <div class="height60 marginTop5 marginBottom10" data-expected-height="60"/>
  <div data-expected-height="75"/>
</div>

<pre>Item min-height: 60px; &amp; margin-left: 5px; &amp; margin-right: 10px;</pre>

<div class="grid">
  <div class="minHeight60 marginTop5 marginBottom10" data-expected-height="60"/>
  <div data-expected-height="75"/>
</div>

<pre>Item height: 60px; &amp; padding-left: 6px;</pre>

<div class="grid">
  <div class="height60 paddingTop6" data-expected-height="66"/>
  <div data-expected-height="66"/>
</div>

<pre>Item min-height: 60px; &amp; padding-left: 6px;</pre>

<div class="grid">
  <div class="minHeight60 paddingTop6" data-expected-height="66"/>
  <div data-expected-height="66"/>
</div>

<pre>Item height: 60px; &amp; padding-right: 3px;</pre>

<div class="grid">
  <div class="height60 paddingBottom3" data-expected-height="63"/>
  <div data-expected-height="63"/>
</div>

<pre>Item min-height: 60px; &amp; padding-right: 3px;</pre>

<div class="grid">
  <div class="minHeight60 paddingBottom3" data-expected-height="63"/>
  <div data-expected-height="63"/>
</div>

<pre>Item height: 60px; &amp; padding-left: 6px; &amp; padding-right: 3px;</pre>

<div class="grid">
  <div class="height60 paddingTop6 paddingBottom3" data-expected-height="69"/>
  <div data-expected-height="69"/>
</div>

<pre>Item min-height: 60px; &amp; padding-left: 6px; &amp; padding-right: 3px;</pre>

<div class="grid">
  <div class="minHeight60 paddingTop6 paddingBottom3" data-expected-height="69"/>
  <div data-expected-height="69"/>
</div>

<pre>Item height: 60px; &amp; border-left-width: 2px;</pre>

<div class="grid">
  <div class="height60 borderTop2" data-expected-height="62"/>
  <div data-expected-height="62"/>
</div>

<pre>Item min-height: 60px; &amp; border-left-width: 2px;</pre>

<div class="grid">
  <div class="minHeight60 borderTop2" data-expected-height="62"/>
  <div data-expected-height="62"/>
</div>

<pre>Item height: 60px; &amp; border-right-width: 4px;</pre>

<div class="grid">
  <div class="height60 borderBottom4" data-expected-height="64"/>
  <div data-expected-height="64"/>
</div>

<pre>Item min-height: 60px; &amp; border-right-width: 4px;</pre>

<div class="grid">
  <div class="minHeight60 borderBottom4" data-expected-height="64"/>
  <div data-expected-height="64"/>
</div>

<pre>Item height: 60px; &amp; border-left-width: 2px; &amp; border-right-width: 4px;</pre>

<div class="grid">
  <div class="height60 borderTop2 borderBottom4" data-expected-height="66"/>
  <div data-expected-height="66"/>
</div>

<pre>Item min-height: 60px; &amp; border-left-width: 2px; &amp; border-right-width: 4px;</pre>

<div class="grid">
  <div class="minHeight60 borderTop2 borderBottom4" data-expected-height="66"/>
  <div data-expected-height="66"/>
</div>

<pre>Item height: 60px; &amp; margin-left: 5px; &amp; margin-right: 10px; &amp; padding-left: 6px; &amp; padding-right: 3px; &amp; border-left-width: 2px; &amp; border-right-width: 4px;</pre>

<div class="grid">
  <div class="height60 marginTop5 marginBottom10 paddingTop6 paddingBottom3 borderTop2 borderBottom4" data-expected-height="75"/>
  <div data-expected-height="90"/>
</div>

<pre>Item min-height: 60px; &amp; margin-left: 5px; &amp; margin-right: 10px; &amp; padding-left: 6px; &amp; padding-right: 3px; &amp; border-left-width: 2px; &amp; border-right-width: 4px;</pre>

<div class="grid">
  <div class="minHeight60 marginTop5 marginBottom10 paddingTop6 paddingBottom3 borderTop2 borderBottom4" data-expected-height="75"/>
  <div data-expected-height="90"/>
</div>

<h3>Direction RTL</h3>

<pre>Item height: 60px;</pre>

<div class="grid directionRTL">
  <div class="height60" data-expected-height="60"/>
  <div data-expected-height="60"/>
</div>

<pre>Item min-height: 60px;</pre>

<div class="grid directionRTL">
  <div class="minHeight60" data-expected-height="60"/>
  <div data-expected-height="60"/>
</div>

<pre>Item height: 60px; &amp; margin-left: 5px;</pre>

<div class="grid directionRTL">
  <div class="height60 marginTop5" data-expected-height="60"/>
  <div data-expected-height="65"/>
</div>

<pre>Item min-height: 60px; &amp; margin-left: 5px;</pre>

<div class="grid directionRTL">
  <div class="minHeight60 marginTop5" data-expected-height="60"/>
  <div data-expected-height="65"/>
</div>

<pre>Item min-height: 60px; &amp; margin-top: 5px; &amp; margin-left: auto</pre>

<div class="grid directionRTL">
  <div class="minHeight60 marginTop5 marginLeftAuto" data-expected-height="60"/>
  <div data-expected-height="65"/>
</div>

<pre>Item height: 60px; &amp; margin-right: 10px;</pre>

<div class="grid directionRTL">
  <div class="height60 marginBottom10" data-expected-height="60"/>
  <div data-expected-height="70"/>
</div>

<pre>Item min-height: 60px; &amp; margin-right: 10px;</pre>

<div class="grid directionRTL">
  <div class="minHeight60 marginBottom10" data-expected-height="60"/>
  <div data-expected-height="70"/>
</div>

<pre>Item height: 60px; &amp; margin-left: 5px; &amp; margin-right: 10px;</pre>

<div class="grid directionRTL">
  <div class="height60 marginTop5 marginBottom10" data-expected-height="60"/>
  <div data-expected-height="75"/>
</div>

<pre>Item min-height: 60px; &amp; margin-left: 5px; &amp; margin-right: 10px;</pre>

<div class="grid directionRTL">
  <div class="minHeight60 marginTop5 marginBottom10" data-expected-height="60"/>
  <div data-expected-height="75"/>
</div>

<pre>Item height: 60px; &amp; padding-left: 6px;</pre>

<div class="grid directionRTL">
  <div class="height60 paddingTop6" data-expected-height="66"/>
  <div data-expected-height="66"/>
</div>

<pre>Item min-height: 60px; &amp; padding-left: 6px;</pre>

<div class="grid directionRTL">
  <div class="minHeight60 paddingTop6" data-expected-height="66"/>
  <div data-expected-height="66"/>
</div>

<pre>Item height: 60px; &amp; padding-right: 3px;</pre>

<div class="grid directionRTL">
  <div class="height60 paddingBottom3" data-expected-height="63"/>
  <div data-expected-height="63"/>
</div>

<pre>Item min-height: 60px; &amp; padding-right: 3px;</pre>

<div class="grid directionRTL">
  <div class="minHeight60 paddingBottom3" data-expected-height="63"/>
  <div data-expected-height="63"/>
</div>

<pre>Item height: 60px; &amp; padding-left: 6px; &amp; padding-right: 3px;</pre>

<div class="grid directionRTL">
  <div class="height60 paddingTop6 paddingBottom3" data-expected-height="69"/>
  <div data-expected-height="69"/>
</div>

<pre>Item min-height: 60px; &amp; padding-left: 6px; &amp; padding-right: 3px;</pre>

<div class="grid directionRTL">
  <div class="minHeight60 paddingTop6 paddingBottom3" data-expected-height="69"/>
  <div data-expected-height="69"/>
</div>

<pre>Item height: 60px; &amp; border-left-width: 2px;</pre>

<div class="grid directionRTL">
  <div class="height60 borderTop2" data-expected-height="62"/>
  <div data-expected-height="62"/>
</div>

<pre>Item min-height: 60px; &amp; border-left-width: 2px;</pre>

<div class="grid directionRTL">
  <div class="minHeight60 borderTop2" data-expected-height="62"/>
  <div data-expected-height="62"/>
</div>

<pre>Item height: 60px; &amp; border-right-width: 4px;</pre>

<div class="grid directionRTL">
  <div class="height60 borderBottom4" data-expected-height="64"/>
  <div data-expected-height="64"/>
</div>

<pre>Item min-height: 60px; &amp; border-right-width: 4px;</pre>

<div class="grid directionRTL">
  <div class="minHeight60 borderBottom4" data-expected-height="64"/>
  <div data-expected-height="64"/>
</div>

<pre>Item height: 60px; &amp; border-left-width: 2px; &amp; border-right-width: 4px;</pre>

<div class="grid directionRTL">
  <div class="height60 borderTop2 borderBottom4" data-expected-height="66"/>
  <div data-expected-height="66"/>
</div>

<pre>Item min-height: 60px; &amp; border-left-width: 2px; &amp; border-right-width: 4px;</pre>

<div class="grid directionRTL">
  <div class="minHeight60 borderTop2 borderBottom4" data-expected-height="66"/>
  <div data-expected-height="66"/>
</div>

<pre>Item height: 60px; &amp; margin-left: 5px; &amp; margin-right: 10px; &amp; padding-left: 6px; &amp; padding-right: 3px; &amp; border-left-width: 2px; &amp; border-right-width: 4px;</pre>

<div class="grid directionRTL">
  <div class="height60 marginTop5 marginBottom10 paddingTop6 paddingBottom3 borderTop2 borderBottom4" data-expected-height="75"/>
  <div data-expected-height="90"/>
</div>

<pre>Item min-height: 60px; &amp; margin-left: 5px; &amp; margin-right: 10px; &amp; padding-left: 6px; &amp; padding-right: 3px; &amp; border-left-width: 2px; &amp; border-right-width: 4px;</pre>

<div class="grid directionRTL">
  <div class="minHeight60 marginTop5 marginBottom10 paddingTop6 paddingBottom3 borderTop2 borderBottom4" data-expected-height="75"/>
  <div data-expected-height="90"/>
</div>
</body></html>
